Modifying plugin for more singular purpose
We are modifying this plugin to more cleanly integrate with
our environment where we do want to automatically create
repositories, but consider this a seperate operation unrelated
to gitosis. This is the first commit towards seperating the
functionality.

The repository creation mechanisms have been sequestered into
a plugin dubbed redmine_autorepo.